+++ httomo/method_wrappers/__init__.py
@@ -1,0 +1,72 @@
+from httomo.runner.method_wrapper import MethodWrapper
+from httomo.runner.methods_repository_interface import MethodRepository
+
+from .generic import GenericMethodWrapper
+
+# import all other wrappers to make sure they are available to the factory function
+# (add imports here when createing new wrappers)
+import httomo.method_wrappers.dezinging  # noqa: F401; pylint: disable=unused-variable
+import httomo.method_wrappers.images  # noqa: F401; pylint: disable=unused-variable
+import httomo.method_wrappers.reconstruction  # noqa: F401; pylint: disable=unused-variable
+import httomo.method_wrappers.rotation  # noqa: F401; pylint: disable=unused-variable
+import httomo.method_wrappers.stats_calc  # noqa: F401; pylint: disable=unused-variable
+
+from mpi4py.MPI import Comm
+from typing import Dict
+
+
+def make_method_wrapper(
+    method_repository: MethodRepository,
+    module_path: str,
+    method_name: str,
+    comm: Comm,
+    output_mapping: Dict[str, str] = {},
+    **kwargs,
+) -> MethodWrapper:
+    """Factory function to generate the appropriate wrapper based on the module
+    path and method name. Clients do not need to be concerned about which particular
+    derived class is returned.
+
+    Parameters
+    ----------
+
+    method_repository: MethodRepository
+        Repository of methods that we can use the query properties
+    module_path: str
+        Path to the module where the method is in python notation, e.g. "httomolibgpu.prep.normalize"
+    method_name: str
+        Name of the method (function within the given module)
+    comm: Comm
+        MPI communicator object
+    output_mapping: Dict[str, str]
+        A dictionary mapping output names to translated ones. The side outputs will be renamed
+        as specified, if the parameter is given. If not, no side outputs will be passed on.
+    kwargs:
+        Arbitrary keyword arguments that get passed to the method as parameters.
+
+    Returns
+    -------
+
+    MethodWrapper
+        An instance of a wrapper class
+    """
+
+    # go throw all subclasses of GenericMethodWrapper and see which one should be instantiated,
+    # based on module path and method name
+    cls: type[GenericMethodWrapper] = GenericMethodWrapper
+    for c in GenericMethodWrapper.__subclasses__():
+        if c.should_select_this_class(module_path, method_name):
+            if cls != GenericMethodWrapper:
+                raise ValueError(
+                    f"The values returned from should_select_this_class('{module_path}', '{method_name}')"
+                    + f" are ambigious between {c.__name__} and {cls.__name__}"
+                )
+            cls = c
+    return cls(
+        method_repository=method_repository,
+        module_path=module_path,
+        method_name=method_name,
+        comm=comm,
+        output_mapping=output_mapping,
+        **kwargs,
+    )
