+++ httomo/runner/platform_section.py
@@ -3,7 +3,7 @@ from typing import Iterator, List
 import mpi4py
 from httomo.runner.pipeline import Pipeline
 from httomo.utils import Colour, Pattern, log_once
-from httomo.runner.backend_wrapper import BackendWrapper
+from httomo.runner.method_wrapper import MethodWrapper
 
 
 class PlatformSection:
@@ -16,7 +16,7 @@ class PlatformSection:
         pattern: Pattern,
         reslice: bool,
         max_slices: int,
-        methods: List[BackendWrapper],
+        methods: List[MethodWrapper],
         save_result: bool = False,
         is_last: bool = False
     ):
@@ -28,13 +28,13 @@ class PlatformSection:
         self.save_result = save_result
         self.is_last = is_last
 
-    def __iter__(self) -> Iterator[BackendWrapper]:
+    def __iter__(self) -> Iterator[MethodWrapper]:
         return iter(self.methods)
 
     def __len__(self) -> int:
         return len(self.methods)
 
-    def __getitem__(self, idx: int) -> BackendWrapper:
+    def __getitem__(self, idx: int) -> MethodWrapper:
         return self.methods[idx]
 
 
@@ -43,7 +43,7 @@ def sectionize(pipeline: Pipeline, save_all: bool = False) -> List[PlatformSecti
 
     # The functions below are internal to reduce duplication
 
-    def needs_global_input(method: BackendWrapper) -> bool:
+    def needs_global_input(method: MethodWrapper) -> bool:
         return method.config_params.get("glob_stats", False)
 
     def is_pattern_compatible(a: Pattern, b: Pattern) -> bool:
@@ -52,7 +52,7 @@ def sectionize(pipeline: Pipeline, save_all: bool = False) -> List[PlatformSecti
     # loop carried variables, to build up the sections
     current_gpu: bool = False
     current_pattern: Pattern = pipeline.loader_pattern
-    current_methods: List[BackendWrapper] = []
+    current_methods: List[MethodWrapper] = []
     save_previous_result: bool = False
 
     def finish_section(needs_reslice=False, save_previous_result=False):
